notes-app/notes-backend/requirements.txt
Flask>=1.0.0
google-cloud-datastore==1.7.1
gunicorn==19.9.0
Flask-SQLAlchemy==2.3.2
PyMySQL==0.9.2
Flask-PyMongo>=2.0.0
oauth2client==4.1.2
PyMongo==3.7.2
six==1.11.0
flask-cors>=1.0.0
requests>=2.22.0
google-cloud-pubsub==0.39.1